scripts/requirements.txt
astroid==2.2.5
colorama==0.4.1
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
Pillow==6.0.0
pylint==2.3.1
six==1.12.0
typed-ast==1.3.5
wrapt==1.11.1
